requirements.txt
Pillow
freetype-py

# Web UI
fastapi
uvicorn[standard]
jinja2
pydantic
httpx

# OCR engines / wrappers
pytesseract
easyocr
transformers
accelerate
torch
torchvision

# PaddleOCR (requires paddlepaddle separately; see Dockerfile / README)
# paddleocr
